[FEAT] listener Interface 추출
